+++ src/main/java/com/ict/mapper/UserInfoMapper.java
@@ -28,7 +28,7 @@ public interface UserInfoMapper {
      * @param user_id 用户id
      * @return
      */
-    UserInfo selectUserInfoByUserId(@Param("user_id") Integer user_id);
+    UserInfo selectUserInfoByUserId(@Param("user_id") Long user_id);
 
     /**
      * 增加用户作品数
